Documents/Gauntlet/ZappierTriggersAPI/backend/requirements.txt
# Zapier Triggers API - Python Dependencies

# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Redis
redis==5.0.1
aioredis==2.0.1

# AWS
boto3==1.34.34
aioboto3==12.3.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Server-Sent Events
sse-starlette==2.0.0

# Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Utilities
python-multipart==0.0.9
python-dateutil==2.8.2
orjson==3.9.13
ulid-py==1.1.0

# Observability
structlog==24.1.0
prometheus-client==0.19.0

# OpenTelemetry - Tracing
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-propagators-aws==1.0.1

# OpenTelemetry - Instrumentation
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-httpx==0.43b0
opentelemetry-instrumentation-sqlalchemy==0.43b0
opentelemetry-instrumentation-redis==0.43b0

# Development
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0
factory-boy==3.3.0
faker==22.5.1

# Linting & Formatting
ruff==0.2.1
black==24.1.1
mypy==1.8.0
types-python-dateutil==2.8.19.20240106
types-redis==4.6.0.20240106

# Load Testing
locust==2.23.1
